feat: provide `client` command
This command downloads and installs the latest version of the `safe` binary.

The Github API is queried for the latest release, from which we can get the version number for
either the client or the node. We then check to see if we have an asset for the current platform,
and if so, we then proceed to download the asset from S3. We download the tar.gz and extract it to
the specified location.

The default install location varies depending on whether the install runs as an elevated user or
not.

The process will be the exact same for the `node` command, but I haven't implemented this yet
because the node binary needs to be renamed.

There are three components involved that been tested separately. They are using `httpmock` to avoid
hitting either the Github API or S3 during the tests. This does involve a bit of setup which can be
a bit tedious, but I do believe it will be worth it to try ensuring we don't break the process. I
will be adding integration tests in CI rather than defining them in Rust.

Modifying the Bash profile has not been implemented yet and Windows also needs some work.