feat: modify shell profile for safe on path
If not running in elevated mode, the shell profile will be modified to add an extra statement to
source an `env` file, in the same manner as `rustup` operates. Doing it this way makes it easier to
modify the profile only once.

Right now this only works for a Bash shell. Cross platform considerations will follow shortly.