feat: provide node command
The only real modification required for this was to the `env` file to also put `~/.safe/node` on the
PATH. Otherwise it goes through identical code paths to the client installation, which has already
been tested quite thoroughly.

Strictly speaking, the `install_bin` function should have test coverage for the node, but as I
already explained previously, the setup code for these tests is quite involved, so there would be a
lot of duplication with minimal benefit.