ci: run some integration tests
A few integration tests were defined in CI. They were done here because it's just a bit easier to
use new shell sessions than if they were defined as Rust tests. The tests install a specific version
of the binaries, which will avoid rate limiting because it doesn't use the Github API. One of the
tests runs non-elevated, which then uses a new shell session to make sure safe is available, i.e.,
that the change to the environment variable has been applied. The other test runs in non-elevated
mode and checks that the binaries are immediately available. On Github Actions, in macOS you can
only run elevated, so there is no non-elevated test.

The usage of `reqwest` was changed to switch off default features, which prevents the `openssl`
crate being referenced in the dependency graph. We have this crate denied in our shared `cargo deny`
configuration, although I don't know why. The `rustls-tls` feature then needs to be explicitly
enabled for using HTTPS.

Since this was the first time Clippy ran in CI, some flagged issues were fixed.