feat: full installations of binaries on windows
For installation on Windows, a function for checking prerequisites was introduced, and the
`configure_shell_profile` function was implemented.

The prerequisite check tests whether the Visual C++ Redistributable is installed by checking the
registry. It would have been possible to install it if it was missing, but you run into
complications there with the need for elevated privileges. Instead, we simply just suggest to the
user that they download and install it. On Linux/macOS, this function is left empty because no
prerequisites are required.

The `configure_shell_profile` function puts the location of `safe` on the user Path variable. To
make the change permanent, it needs to be done by changing a registry value.

Test coverage was not provided for either of these functions because it would involve mocking out
the registry usage, and that didn't really seem worth it.