refactor: accommodate multiple github releases
BREAKING CHANGE: the latest version of a binary is obtained by querying all the releases on the
`safe_network` repository. The previous mechanism used the 'latest' release, which no longer works.

Our previous release process used a single Github Release to distribute the three binaries. To
determine what was the latest version of any of the binaries, we could just get the 'latest'
release, which is an API Github make available.

Now we use separate releases for each binary. We therefore have to get a list of all the releases,
then determine which is the latest for either of the three binaries. We then pull that specific
release and get the name of the asset to be pulled from S3.