feat: use `~/.local/bin` as install location
BREAKING CHANGE: rather than use `~/.safe/cli` and `~/.safe/node`, we can put all the binaries in
`~/.local/bin`, which is part of the XDG Base Directory Specification. This does not change the
install location for when we're running as the root user.

On Windows, the install location is changed slightly, to use `%USERPROFILE%\safe`, rather than
`%USERPROFILE%\.safe` since the use of `.` prefixed directories are not as common on Windows and
wouldn't have them marked as hidden by software like Windows Explorer. There is no common location
on Windows to put binaries without having Administrator access.

The integration tests are updated to take this new location into account, and the version numbers of
the components are changed to use up-to-date numbers from releases in the new repository.