feat: provide unix-based install script
This script detects the OS, then CPU architecture, then obtains the latest version of `safeup`. With
that information, it downloads the relevant `safeup` binary, extracts it, then uses it to install
the client by default.

The script is intended to be compatible with both Linux and macOS, but also x86_64 and ARM CPU
architectures.

A modification is also made to safeup to inform the user that as an alternative to sourcing the `env`
file to make the modified `PATH` available, they can also start a new session.