feat: do not install client by default
BREAKING CHANGE: for the Unix installation, the default client install is removed.

I quickly realised this behaviour was quite silly, because you might run the installer on a node
machine and have no interest in the client.

As an alternative, `--client` and `--node` arguments were added to the install script to support
installing either of these without an additional post-install step.

They have been left out of the Windows install script because Powershell does not support sending
the arguments when the script is downloaded, like you can with Bash.