feat: provide `update` command
Iterates over the installed components and updates them if a later version is available.

This resulted in some shared code being refactored and the `Settings` struct being updated to keep
track of the components and versions that were installed.

The text on the original install commands were also improved for clarity.