refactor: download and extract using `sn-releases` crate
The code for downloading and extracting release binaries from Github and S3 was being repeated in
several different places, so we decided to put some common code for it in a crate.

We are now updating `safeup` to use it, which removes both the `github` and `s3` modules, and makes
the `install_bin` tests simpler.

From the user's point of view, functionally everything should be the same.

I also added a quick Vagrantfile here that can be used for testing safeup. This is useful if you
don't want your local machine littered with binaries that were only used for testing.