feat: provide `node-manager` command
BREAKING CHANGE: the new settings file will have additional entries for keeping track of the
safenode-manager installation. Previously serialised settings files will be incompatible with this
change. Users will need to clear their previous settings file when they upgrade.

This will install the node manager alongside the other binaries. After the `sn-releases` crate was
extended for the new release type, adding the new command was quite straight forward, since it's
just the installation of another binary.

At some point later we may need to extend this installation to also distribute the RPC client, which
may be required for the node manager if we're going to use it in service shutdown commands. For now
though, it's just distributing the `safenode-manager` binary.