chore: remove `testnet` command
We have removed support for the `testnet` binary, so it no longer applies and is hence being
removed.

Also upgrades `sn-releases` to the latest version, which uses the crates.io API to get the latest
version of a crate.

BREAKING CHANGE: properties have been removed from the `Settings` struct, meaning settings files
from previous versions will be incompatible with this new version.