refactor: use version type rather than strings
Upgrade to `sn-releases` version 0.2.0, which forces the use of the `semver::Version` type rather
than `String` for dealing with versions.

It made sense to do the same thing with string-based version fields in this crate. All of the fields
on the `Settings` type were also turned into an `Option`, which makes sense because it's possible
that we don't have one of the asset types installed. I'm not sure why this wasn't done in the first
place. It made the code a bit simpler.

BREAKING CHANGE: settings files from previous versions will not deserialize properly and will have
to be deleted.