feat: `client` cmd now uses the `autonomi` binary
The `client` command will now retrieve and install the `autonomi` binary rather than `safe`.

The installer scripts have also been updated to use the new settings locations.

BREAKING CHANGE: the binaries and settings now use a path with `autonomi` rather than `safe`. The
old settings file is now also incompatible because it is at a different location and the field names
have changed, so they will serialize/deserialize to/from different values. Fetching a different
binary itself is also a breaking change.

The tests in the PR workflow are updated to work with the newer versions of the `safenode` and
`safenode-manager` binaries, which have new output for `--version`. We need to use RC versions just
now because those are the only ones that have been released with the new versioning information.